// src/app/models/address-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface AddressView {
    id?: number | string;
    line1?: string;
    line2?: string;
    line3?: string;
    city?: string;
    stateId?: number | string;
    pincode?: string;
    countryId?: number | string;
}