// src/app/models/contact-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface ContactView {
    id?: number | string;
    website?: string;
    email1?: string;
    email2?: string;
    mobileTel?: string;
    landTel?: string;
    otherTel?: string;
}