// src/app/modules/lshp-admin/models/post-user-view-model.ts
import { UserPersonViewModel } from './user-person-view-model';

/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Gets or Sets
 */
export interface PostUserViewModel { 

    instituteId?: string;
    /**
     * Gets or sets
     */
    userName?: string;
    /**
     * Gets or sets
     */
    firstName?: string;
    /**
     * Gets or sets
     */
    lastName?: string;
    /**
     * Gets or sets
     */
    middleName?: string;
    /**
     * Gets or sets
     */
    genderTypeId?: number;
    /**
     * Gets or sets
     */
    dateOfBirth?: string;

    /**
     * Gets or sets
     */
    person?: UserPersonViewModel;
}