// src/app/models/user-management/institute-user-view.ts
import { UserPersonViewModel } from 'app/modules/lshp-admin/models/user-person-view-model';

/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



/**
 * 
 */
export interface InstituteUserViewModel {
    /**
     * Gets or Sets Id
     */
    id?: string;
    /**
     * Gets or sets
     */
    firstName?: string;
    /**
     * Gets or sets
     */
    lastName?: string;
    /**
     * Gets or sets
     */
    middleName?: string;
    /**
     * Gets or sets
     */
    dateOfBirth?: Date;
    /**
     * Gets or sets
     */
    person?: UserPersonViewModel;

    /**
    * Gets or Sets InstituteId
    */
    instituteId?: string;
}